+++ front/src/UI/atoms/index.tsx
@@ -1,7 +1,4 @@
-import { AlarmMessage } from './AlarmMessage';
-import { AlarmUser } from './AlarmUser';
-import { AlarmChatMessage } from './AlarmChatMessage';
 import { ColorModeSwitcher } from './ColorModeSwitcher';
 import { NavIcon } from './NavIcon';
 
-export { AlarmMessage, AlarmUser, AlarmChatMessage, ColorModeSwitcher, NavIcon };
+export { ColorModeSwitcher, NavIcon };
