+++ front/src/UI/molecules/Alarm/index.tsx
@@ -1,0 +1,5 @@
+import { AlarmMessage } from './AlarmMessage';
+import { AlarmUser } from './AlarmUser';
+import { AlarmChatMessage } from './AlarmChatMessage';
+
+export { AlarmMessage, AlarmUser, AlarmChatMessage };
