+++ front/src/UI/organisms/AlarmNotifier/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { AccordionItem, AccordionButton, AccordionPanel, AccordionIcon, Box, Text, Flex } from '@chakra-ui/react';
 
-import { AlarmMessage } from '../../atoms/AlarmMessage';
+import { AlarmMessage } from '../../molecules';
 
 const messageData = [
   {
