+++ front/src/utils/interface.ts
@@ -1,0 +1,9 @@
+/*
+ ** Alarm
+ */
+
+export interface IalarmMessage {
+  title: string;
+  content: string;
+  alarmTime: string;
+}
